Move test_util.py and testdata/ into tests/
