+++ acme/tests/standalone_test.py
@@ -11,9 +11,9 @@ import mock
 import requests
 
 from acme import challenges
-from acme import test_util
 from acme.magic_typing import Set # pylint: disable=unused-import, no-name-in-module
 
+import test_util
 
 class TLSServerTest(unittest.TestCase):
     """Tests for acme.standalone.TLSServer."""
